# Civic-Sentinel-System/lib/api-zod/src/generated/types/agentResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CivicPulse AI - Multi-Agent Civic Issue Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { AgentResultOutput } from './agentResultOutput';
import type { AgentResultStatus } from './agentResultStatus';

export interface AgentResult {
  agentName: string;
  status: AgentResultStatus;
  durationMs: number;
  output?: AgentResultOutput;
  /** @nullable */
  error?: string | null;
}