# Civic-Sentinel-System/lib/api-zod/src/generated/types/resolutionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CivicPulse AI - Multi-Agent Civic Issue Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */

export interface ResolutionInput {
  resolutionNote: string;
  /** @nullable */
  resolvedImageBase64?: string | null;
  /** @nullable */
  resolvedImageUrl?: string | null;
}